Move core schema to code from EDN
This is a step backwards from runtime configurability. But, if this part
changes, it's a really huge deal. So it truly should be painful.

Besides, I want to use this part in other projects. That's at least half
the point to this library. The EDN version wasn't being distributed as
part of the .jar.